Fix bug prone code reported by clang-tidy

+++ vpr/src/pack/pack_report.cpp
@@ -31,8 +31,8 @@ void report_packing_pin_usage(std::ostream& os, const VprContext& ctx) {
     for (auto blk : cluster_ctx.clb_nlist.blocks()) {
         t_logical_block_type_ptr type = cluster_ctx.clb_nlist.block_type(blk);
 
-        inputs_used[type].push_back(cluster_ctx.clb_nlist.block_input_pins(blk).size() + cluster_ctx.clb_nlist.block_clock_pins(blk).size());
-        outputs_used[type].push_back(cluster_ctx.clb_nlist.block_output_pins(blk).size());
+        inputs_used[type].push_back(static_cast<float>(cluster_ctx.clb_nlist.block_input_pins(blk).size() + cluster_ctx.clb_nlist.block_clock_pins(blk).size()));
+        outputs_used[type].push_back(static_cast<float>(cluster_ctx.clb_nlist.block_output_pins(blk).size()));
     }
 
     vtr::OsFormatGuard os_guard(os);
@@ -44,27 +44,27 @@ void report_packing_pin_usage(std::ostream& os, const VprContext& ctx) {
         if (is_empty_type(type)) continue;
         if (!inputs_used.count(type)) continue;
 
-        float max_inputs = *std::max_element(inputs_used[type].begin(), inputs_used[type].end());
-        float min_inputs = *std::min_element(inputs_used[type].begin(), inputs_used[type].end());
-        float avg_inputs = std::accumulate(inputs_used[type].begin(), inputs_used[type].end(), 0) / float(inputs_used[type].size());
+        float max_inputs = static_cast<float>(*std::max_element(inputs_used[type].begin(), inputs_used[type].end()));
+        float min_inputs = static_cast<float>(*std::min_element(inputs_used[type].begin(), inputs_used[type].end()));
+        float avg_inputs = static_cast<float>(std::accumulate(inputs_used[type].begin(), inputs_used[type].end(), 0.f)) / static_cast<float>(inputs_used[type].size());
 
         float max_outputs = *std::max_element(outputs_used[type].begin(), outputs_used[type].end());
         float min_outputs = *std::min_element(outputs_used[type].begin(), outputs_used[type].end());
-        float avg_outputs = std::accumulate(outputs_used[type].begin(), outputs_used[type].end(), 0) / float(outputs_used[type].size());
+        float avg_outputs = std::accumulate(outputs_used[type].begin(), outputs_used[type].end(), 0.f) / static_cast<float>(outputs_used[type].size());
 
         os << "Type: " << type->name << "\n";
 
         os << "\tInput Pin Usage:\n";
-        os << "\t\tMax: " << max_inputs << " (" << max_inputs / float(total_input_pins[type]) << ")"
+        os << "\t\tMax: " << max_inputs << " (" << max_inputs / static_cast<float>(total_input_pins[type]) << ")"
            << "\n";
-        os << "\t\tAvg: " << avg_inputs << " (" << avg_inputs / float(total_input_pins[type]) << ")"
+        os << "\t\tAvg: " << avg_inputs << " (" << avg_inputs / static_cast<float>(total_input_pins[type]) << ")"
            << "\n";
-        os << "\t\tMin: " << min_inputs << " (" << min_inputs / float(total_input_pins[type]) << ")"
+        os << "\t\tMin: " << min_inputs << " (" << min_inputs / static_cast<float>(total_input_pins[type]) << ")"
            << "\n";
 
         if (total_input_pins[type] != 0) {
             os << "\t\tHistogram:\n";
-            auto input_histogram = build_histogram(inputs_used[type], 10, 0, total_input_pins[type]);
+            auto input_histogram = build_histogram(inputs_used[type], 10, 0, static_cast<float>(total_input_pins[type]));
             for (const std::string& line : format_histogram(input_histogram)) {
                 os << "\t\t" << line << "\n";
             }
@@ -81,7 +81,7 @@ void report_packing_pin_usage(std::ostream& os, const VprContext& ctx) {
         if (total_output_pins[type] != 0) {
             os << "\t\tHistogram:\n";
 
-            auto output_histogram = build_histogram(outputs_used[type], 10, 0, total_output_pins[type]);
+            auto output_histogram = build_histogram(outputs_used[type], 10, 0, static_cast<float>(total_output_pins[type]));
             for (auto line : format_histogram(output_histogram)) {
                 os << "\t\t" << line << "\n";
             }
